Rename some index files where foundry has corresponding module.mjs files
