+++ src/common/abstract/backend.d.ts
@@ -1,0 +1,8 @@
+declare namespace foundry {
+  namespace abstract {
+    /**
+     * An interface which is required by both client and server-side to provide implementations for document operations.
+     */
+    abstract class DatabaseBackend {}
+  }
+}
