+++ src/common/abstract/document.d.ts
@@ -1,0 +1,8 @@
+declare namespace foundry {
+  namespace abstract {
+    /**
+     * The abstract base interface for all Document types.
+     */
+    abstract class Document {}
+  }
+}
