+++ src/common/abstract/module.d.ts
