+++ src/common/data/fields.d.ts
@@ -1,0 +1,3 @@
+declare namespace foundry {
+  namespace data {}
+}
