+++ src/common/utils/module.d.ts
